feat(agents): graduate sessions to agents/sessions
