+++ .changeset/calm-rivers-session.md
@@ -1,0 +1,5 @@
+---
+"agents": minor
+---
+
+Graduate the Session API to the stable `agents/sessions` entry point. The previous `agents/experimental/memory/session` path remains available as a deprecated compatibility re-export.
