+++ experimental/README.md
@@ -11,4 +11,4 @@ The code here is for exploration, prototyping, and validating patterns that may 
 - **[forever.md](./forever.md)** — Design doc for durable long-running execution. Covers `keepAlive`, `runFiber` (checkpointing, eviction recovery), and `AIChatAgent` chat recovery via `onChatRecovery`. Built into the `Agent` and `AIChatAgent` base classes.
 - **[forever-fibers/](./forever-fibers/)** — Example of `Agent.runFiber()` for durable background work with real-time progress tracking.
 - **[forever-chat/](./forever-chat/)** — Example of `AIChatAgent` with `chatRecovery` for multi-provider chat recovery (Workers AI, OpenAI, Anthropic).
-- **[session-memory](./session-memory/)** — Example of the Session API for conversation history with automatic compaction. Demonstrates `agents/experimental/memory/session` with micro-compaction and LLM-based summarization.
+- **[session-memory](./session-memory/)** — Example of the Session API for conversation history with automatic compaction. Demonstrates `agents/sessions` with micro-compaction and LLM-based summarization.
