+++ experimental/session-multichat/src/server.ts
@@ -12,11 +12,8 @@ import {
   routeAgentRequest,
   type StreamingResponse
 } from "agents";
-import { SessionManager } from "agents/experimental/memory/session";
-import {
-  truncateOlderMessages,
-  createCompactFunction
-} from "agents/experimental/memory/utils";
+import { SessionManager } from "agents/sessions";
+import { truncateOlderMessages, createCompactFunction } from "agents/sessions";
 import type { UIMessage } from "ai";
 import { createWorkersAI } from "workers-ai-provider";
 import {
