+++ packages/agents/src/experimental/memory/utils/index.ts
@@ -1,24 +1,4 @@
-export {
-  estimateStringTokens,
-  estimateMessageTokens,
-  CHARS_PER_TOKEN,
-  WORDS_TOKEN_MULTIPLIER,
-  TOKENS_PER_MESSAGE
-} from "./tokens";
-
-export { truncateOlderMessages, type TruncateOptions } from "./compaction";
-
-export {
-  createCompactFunction,
-  isCompactionMessage,
-  COMPACTION_PREFIX,
-  type CompactResult,
-  sanitizeToolPairs,
-  alignBoundaryForward,
-  alignBoundaryBackward,
-  findTailCutByTokens,
-  computeSummaryBudget,
-  buildSummaryPrompt,
-  type CompactOptions,
-  type CompactTokenCounter
-} from "./compaction-helpers";
+/**
+ * @deprecated Import Session utilities from `agents/sessions` instead.
+ */
+export * from "../../../sessions/utils";
