+++ packages/agents/src/sessions/search.ts
