+++ packages/agents/src/sessions/types.ts
