+++ packages/agents/src/sessions/utils/compaction.ts
@@ -7,8 +7,8 @@
  * Does NOT mutate stored messages — operates on a copy.
  */
 
-import { truncateToolOutput } from "../../../chat/tool-output-truncation";
-import type { SessionMessage } from "../session/types";
+import { truncateToolOutput } from "../../chat/tool-output-truncation";
+import type { SessionMessage } from "../types";
 
 export interface TruncateOptions {
   /** Number of recent messages to keep intact (default: 4) */
