+++ packages/think/src/tests/agents/assistant-agent-loop.ts
@@ -8,7 +8,7 @@
 import type { LanguageModel, ToolSet, UIMessage } from "ai";
 import { tool } from "ai";
 import { z } from "zod";
-import { Session } from "agents/experimental/memory/session";
+import { Session } from "agents/sessions";
 import type { ObservabilityEvent } from "agents/observability";
 import { Think } from "../../think";
 import type {
